feat: 잘못 설정한 테스트 설정 변경
