+++ README.md
@@ -101,25 +101,20 @@ npm run dev
        - 최대  VUser: 30(VUser) = 31.8 (RPS) *  2.8(T) / 3 (R)
 
 2. Smoke, Load, Stress 테스트 스크립트와 결과를 공유해주세요
-   1. Smoke
-      - 테스트 기간
-          - 총 테스트 시간: 1시간
-          - 평균 VUser : 40분
-          - 최대 VUser : 20분
-      - 테스트 시나리오
+    - 테스트 시나리오
         1. 메인 페이지 접속
         2. 로그인 페이지 접속
         3. 로그인 한번 실패
         4. 로그인 성공
         5. 경로 검색 페이지 접속
         6. 경로 검색 버튼 클릭
+   1. Smoke
+      -  
+       
    2. Stress
-      1. 허용 범위를 초과하여 한계치를 측정하는 테스트
-         1. MaxVUser: 90: 최대 트래픽 시간 VUser 에 3배로 테스트 진행
-         2. Stage: 20 -> 30 -> 50 -> 60 -> 70 -> 80 -> 90 -> 60 -> 30 -> 10
+   
    3. Load
-      1. 부하가 걸린 상황에서 시스템의 동작을 점검
-
+       - 
 ---
 
 ### 3단계 - 로깅, 모니터링
