+++ webTestApp/load.js
@@ -1,40 +1,42 @@
 import http from 'k6/http';
 import {check, sleep} from 'k6';
-import {getRandomPathsQuery} from "./getRandomPathsQuery";
+import {Options} from 'k6/options';
+import {getRandomPathsQuery} from "./getRandomPathsQuery.js";
 
 /**
- * 스트레스 테스트는 극한 조건 에서 시스템의 안정성과 신뢰성을 검증 하기 위함이다.
- * 트래픽이 많이 생기는 할인 행사, 기념일 같은 특정날이 을때 자신의 서버가 어떻게 트래픽을 처리하는지 시뮬레이션 확인할수 있다.
+ * 부하 테스트(loadTest)
+ * 평균, 초대 트래픽 같은 적절한 부하를 발생시켜 정상적으로 서버가 동작하는지 확인하여 신뢰성을 얻는다.
  *
- * 스트레스 테스트 목적
- * - 극한 조건에서 시스템이 작동하는 방식.
- * - 사용자 또는 처리량 측면에서 시스템의 최대 용량 확인
- * - 시스템의 한계점과 실패 환경 확인
- * - 스트레스 테스트가 끝난 후 시스템의 자동 복구 여부
+ *
+ * 테스트 실행법
+ * - 0부터 시작해서 천천히 최대 부하까지 올리기
+ * - 최대에서 다시 0까지 천천히 부하를 내린다.
  */
 
 /**
  * @type {Options}
  */
-export const options = {
+export let options = {
   stages: [
-    {duration: '5m', target: 20},
-    {duration: '5m', target: 30},
-    {duration: '3m', target: 50},
-    {duration: '3m', target: 60},
-    {duration: '3m', target: 70},
-    {duration: '3m', target: 80},
-    {duration: '5m', target: 90},
-    {duration: '3m', target: 60},
-    {duration: '5m', target: 30},
-    {duration: '5m', target: 10},
+    { duration: '10m', target: 5 },
+    { duration: '20m', target: 30 },
+    { duration: '10m', target: 0 },
   ],
-}
+  thresholds: {
+    // 상위 90% 는 0.6 초 미만 / 95% 는 0.7초 미만
+    http_req_duration: ['p(90)< 600', 'p(95) < 700' ],
+  },
+};
 
 const BASE_URL = 'https://xn--119-9j6nx7w.xn--h32bi4v.xn--3e0b707e';
 const USERNAME = 'test@test.com';
 const PASSWORD = 'test1234';
 
+const jsonParams = {
+  headers: {
+    'Content-Type': 'application/json',
+  },
+};
 
 export default function () {
   //1. 메인 페이지 접속
@@ -75,7 +77,7 @@ export default function () {
   check(http.get(`${BASE_URL}/paths?source=${randomPathsQuery.sourceId}&target=${randomPathsQuery.targetId}`), {
     '경로 검색 버튼 클릭': (resp) => resp.status === 200 && resp.json('distance') >= 0,
   });
+
+  // 1초 대기
   sleep(1);
-}
-
-
+};
