+++ webTestApp/smoke.js
@@ -1,30 +1,29 @@
 import http from 'k6/http';
 import {check, sleep} from 'k6';
-import {getRandomPathsQuery} from "./getRandomPathsQuery.js";
+import {getRandomPathsQuery} from "./getRandomPathsQuery";
+import {Options} from "k6/options";
 
-export let options = {
-  stages: [
-    // 35분 동안 평균 트래픽 이후 5분 안에 최고 트래픽으로 올라간 후 20분 유지
-    { duration: '1s', target: 5 },
-    { duration: '35m', target: 5,  },
-    { duration: '5m', target: 30 },
-    { duration: '20m', target: 30 },
-  ],
+/**
+ * 스모크 테스트 (smokeTest)
+ * 최소 부하로 기능에 문제가 없는지 확인하거나 테스트 스크립트가 정상 작동하는지 테스트 한다.
+ */
+
+/**
+ * @type {Options}
+ */
+export const options = {
+  vus: 1,
+  duration: '1m',
   thresholds: {
-    // 상위 90% 는 0.6 초 미만 / 95% 는 0.7초 미만
-    http_req_duration: ['p(90)< 600', 'p(95) < 700' ],
-  },
-};
+    http_req_duration: ['p(99)<700']
+  }
+}
+
 
 const BASE_URL = 'https://xn--119-9j6nx7w.xn--h32bi4v.xn--3e0b707e';
 const USERNAME = 'test@test.com';
 const PASSWORD = 'test1234';
 
-const jsonParams = {
-  headers: {
-    'Content-Type': 'application/json',
-  },
-};
 
 export default function () {
   //1. 메인 페이지 접속
@@ -65,7 +64,7 @@ export default function () {
   check(http.get(`${BASE_URL}/paths?source=${randomPathsQuery.sourceId}&target=${randomPathsQuery.targetId}`), {
     '경로 검색 버튼 클릭': (resp) => resp.status === 200 && resp.json('distance') >= 0,
   });
-
-  // 1초 대기
   sleep(1);
-};
+}
+
+
